Face-Emotion-Detector/requirements.txt
tensorflow==2.9.1
protobuf==3.20.*
numpy
Pillow
openai
